+++ Zilon.Core/Zilon.Core/Schemes/IPropEquipSubScheme.cs
@@ -29,19 +29,4 @@ namespace Zilon.Core.Schemes
         /// </summary>
         EquipmentSlotTypes[]? SlotTypes { get; }
     }
-
-    public interface IPropEquipRestrictions
-    {
-        /// <summary>
-        /// Determines rules to keep thing in hands.
-        /// By default, all equipment in hand slot is one-handed.
-        /// </summary>
-        PropHandUsage? PropHandUsage { get; }
-    }
-
-    public enum PropHandUsage
-    {
-        Undefined = 0,
-        TwoHanded = 1
-    }
 }
