+++ Zilon.Core/Zilon.Core/Schemes/PropEquipRestrictions.cs
@@ -1,0 +1,10 @@
+﻿using Newtonsoft.Json;
+
+namespace Zilon.Core.Schemes
+{
+    public sealed class PropEquipRestrictions : IPropEquipRestrictions
+    {
+        [JsonProperty]
+        public PropHandUsage? PropHandUsage { get; private set; }
+    }
+}
