+++ Zilon.Core/Zilon.Core/Schemes/PropHandUsage.cs
@@ -1,0 +1,11 @@
+﻿namespace Zilon.Core.Schemes
+{
+    /// <summary>
+    /// Rule of restrictions according a person hands.
+    /// </summary>
+    public enum PropHandUsage
+    {
+        Undefined = 0,
+        TwoHanded = 1
+    }
+}
